add all routes

+++ backend/src/modules/course/controller.ts
@@ -1,6 +1,5 @@
 import { Request, Response } from "express";
 import { prisma } from "../../config/database";
-import { Course } from "../../config/generated/prisma";
 
 export const getAllCourses = async (req: Request, res: Response) => {
   try {
@@ -13,24 +12,48 @@ export const getAllCourses = async (req: Request, res: Response) => {
 
 export const addCourse = async (req: Request, res: Response) => {
   let course = req.body;
+  /*
+  course = {
+  ...
+  instructorId ,
+  overviews: [ {}, {}, {}...],
+  details: [
+  {
+    week , title 
+    lessons : [ {}, {},{}...]
+  }, 
+  {
+   week , title 
+    lessons : [ {}, {},{}...]
+  } ,
+  {
+   week , title 
+    lessons : [ {}, {},{}...]
+  }...]
+  }
+  */
+
   // validate the course data
 
   try {
     const resultcourse = await prisma.course.create({
       data: {
+        ...course,
+
+        overviews: {
+          create: course.overviews,
+        },
         details: {
           //@ts-ignore
           create: course.details.map((detail) => {
             return { ...detail, lessons: { create: detail.lessons } };
           }),
         },
-        ...course,
-        overviews: {
-          create: course.overviews,
-        },
       },
     });
-    return res.status(201).json({ message: "create new course", data: course });
+    return res
+      .status(201)
+      .json({ message: "create new course", data: resultcourse });
   } catch (error) {
     return res.status(400).json({ error });
   }
@@ -46,30 +69,111 @@ export const getCourseDetails = async (req: Request, res: Response) => {
       include: {
         instructor: true,
         overviews: true,
-        details: true,
+        details: {
+          include: {
+            lessons: true,
+          },
+        },
         Comment: { include: { student: true } },
       },
     });
     return res
-      .status(201)
+      .status(200)
       .json({ message: "get course details", data: course });
   } catch (error) {
     return res.status(400).json({ error });
   }
 };
 
-export const updateCourse = async () => {};
+export const updateCourse = async (req: Request, res: Response) => {
+  try {
+    const { id } = req.params;
+    let course = req.body;
+
+    /*
+  course = {
+  ...
+  instructorId ,
+  overviews: [ {}, {}, {}...],
+  details: [
+  {
+    week , title 
+    lessons : [ {}, {},{}...]
+  }, 
+  {
+   week , title 
+    lessons : [ {}, {},{}...]
+  } ,
+  {
+   week , title 
+    lessons : [ {}, {},{}...]
+  }...]
+  }
+  */
+
+    // validate the course data
+
+    //delete overviews and details
+    await prisma.overViews.deleteMany({
+      where: {
+        courseId: id,
+      },
+    });
+
+    await prisma.details.deleteMany({
+      where: {
+        courseId: id,
+      },
+    });
+
+    //update course
+    const updatedCourse = await prisma.course.update({
+      where: {
+        id,
+      },
+      data: {
+        ...course,
+
+        overviews: {
+          create: course.overviews,
+        },
+        details: {
+          //@ts-ignore
+          create: course.details.map((detail) => {
+            return { ...detail, lessons: { create: detail.lessons } };
+          }),
+        },
+      },
+    });
+    return res
+      .status(200)
+      .json({ message: "course updated OK", updatedCourse });
+  } catch (error) {
+    return res.status(400).json({ error });
+  }
+};
 
 export const deleteCourse = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+
+    const course = await prisma.course.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!course || course.type === "course") {
+      return res.status(404).json({ message: "course not found" });
+    }
+
     const deletedcourse = await prisma.course.delete({
       where: {
         id,
       },
     });
     return res
-      .status(201)
+      .status(200)
       .json({ message: "deleted course OK", data: deletedcourse });
   } catch (error) {
     return res.status(400).json({ error });
